bugbounty_tool_project/requirements.txt
argparse
asyncio
click
jinja2
openai
pyyaml
requests
tqdm
pytest